% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIFloraExplorer-package.R
\docType{package}
\name{BIFloraExplorer-package}
\alias{BIFloraExplorer}
\alias{BIFloraExplorer-package}
\title{BIFloraExplorer: A taxonomic, genetic and ecological data resource for the vascular plants of Britain and Ireland}
\description{
A comprehensive attribute database of all vascular plant species currently extant in Britain and Ireland.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RBGKew/BIFloraExplorer}
}

}
\author{
\strong{Maintainer}: M.C. Henniges \email{marie.c.henniges@gmail.com} (\href{https://orcid.org/0000-0003-3626-2830}{ORCID})

Authors:
\itemize{
  \item R.F. Powell (\href{https://orcid.org/0000-0001-7361-3164}{ORCID})
  \item S. Mian (\href{https://orcid.org/0000-0001-7828-0629}{ORCID})
  \item C.A. Stace
  \item K.J. Walker
  \item R.J. Gornall (\href{https://orcid.org/0000-0002-3814-8751}{ORCID})
  \item M.J.M. Christenhusz (\href{https://orcid.org/0000-0003-1398-8743}{ORCID})
  \item M.R. Brown (\href{https://orcid.org/0000-0003-2561-516X}{ORCID})
  \item A.D. Twyford (\href{https://orcid.org/0000-0002-8746-6617}{ORCID})
  \item P.M. Hollingsworth (\href{https://orcid.org/0000-0003-0602-0654}{ORCID})
  \item L. Jones (\href{https://orcid.org/0000-0002-8197-1970}{ORCID})
  \item N. De Vere (\href{https://orcid.org/0000-0001-9593-6925}{ORCID})
  \item A. Antonelli (\href{https://orcid.org/0000-0003-1842-9297}{ORCID})
  \item A.R. Leitch (\href{https://orcid.org/0000-0001-8574-302X}{ORCID})
  \item I.J. Leitch (\href{https://orcid.org/0000-0002-3837-8186}{ORCID})
}

}
\keyword{internal}
